Table list and permission view
Added action on IndexTool to show a user's tables and the perimissions
on them.
Added drive.metadata.readonly scope to apis on FusionTable to support
the permission view
Added a work around for popping up the default browser to grant
permissions, as I don't use the default browser for Aikuma-related
stuff.