Move metadata validation to Utils
Moved the validation method from FusionIndex to Utils. This allows one
to validate the metadata without having to actually call the
index/update methods. The method also provides a string response with a
complete list of what's wrong with a given metadata map, which is
probably more generall useful than before when it barfed up an
InvalidArgumentException on the first problem.

This is mostly to support the index_server, since it'd be nice to
provide better information about what's wrong with the data (preventing
round trips).